+++ src/lsp/readFileRequest.ts
@@ -1,13 +1,17 @@
 import * as vslc from 'vscode-languageclient';
 
-export const READ_FILE_REQUEST_METHOD = 'sorbet/readFile';
+export const READ_FILE_REQUEST = new vslc.RequestType<
+  vslc.TextDocumentIdentifier,
+  vslc.TextDocumentItem,
+  void
+>('sorbet/readFile');
 
 /**
  * See https://sorbet.org/docs/lsp#sorbetreadfile-request
  */
 export interface ReadFileRequest {
   sendRequest(
-    method: typeof READ_FILE_REQUEST_METHOD,
+    requestType: typeof READ_FILE_REQUEST,
     param: vslc.TextDocumentIdentifier,
     token?: vslc.CancellationToken,
   ): Promise<vslc.TextDocumentItem>;