+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.3.14
 - Add a heuristic `SelectionRangeProvider` to improve behavior of **Expand Selection**, **Shrink Selection** and related commands.
+- Add support for `sorbet/hierarchyReferences` via custom **Peek Usages** context menu command.
 
 ## 0.3.13
 - `sorbet/config` files now use the new **sorbet-config** language.